Use parameterized tests for property containers
- Using parameterized tests, we test all implementations of IPropertyContainerConstTest using the same tests, as they are expected to work the same anyway. The PropertyContainer class still has some of its own tests for the setter methods etc.
- Added a general PropertyTestUtil class to consolidate stuff like creating properties for testing.
