+++ animatedarchitecture-core/src/test/java/nl/pim16aap2/animatedarchitecture/core/structures/properties/PropertyContainerTest.java
@@ -1,7 +1,6 @@
 package nl.pim16aap2.animatedarchitecture.core.structures.properties;
 
 import nl.pim16aap2.animatedarchitecture.core.structures.StructureType;
-import nl.pim16aap2.animatedarchitecture.core.util.Constants;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -12,42 +11,16 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.mockito.junit.jupiter.MockitoSettings;
 import org.mockito.quality.Strictness;
 
-import java.util.List;
 import java.util.Objects;
 import java.util.UUID;
+
+import static nl.pim16aap2.animatedarchitecture.core.structures.properties.PropertyTestUtil.*;
 
 @Timeout(1)
 @ExtendWith(MockitoExtension.class)
 @MockitoSettings(strictness = Strictness.LENIENT)
 class PropertyContainerTest
 {
-    private static final Property<Integer> PROPERTY_UNSET = new Property<>(
-        "external",
-        "unset_property",
-        Integer.class,
-        5
-    );
-
-    private static final String PROPERTY_STRING_DEFAULT = "default";
-    private static final Property<String> PROPERTY_STRING = new Property<>(
-        Constants.PLUGIN_NAME,
-        "string_property",
-        String.class,
-        PROPERTY_STRING_DEFAULT
-    );
-
-    private static final Property<Object> PROPERTY_NULLABLE = new Property<>(
-        Constants.PLUGIN_NAME,
-        "nullable_property",
-        Object.class,
-        null
-    );
-
-    private static final List<Property<?>> PROPERTIES = List.of(
-        PROPERTY_STRING,
-        PROPERTY_NULLABLE
-    );
-
     private PropertyContainer propertyContainer;
 
     @BeforeEach
@@ -56,35 +29,6 @@ class PropertyContainerTest
         propertyContainer = PropertyContainer.forProperties(PROPERTIES);
     }
 
-    @Test
-    void testGetNullablePropertyValue()
-    {
-        final var value = propertyContainer.getPropertyValue(PROPERTY_NULLABLE);
-
-        Assertions.assertNotNull(value);
-        Assertions.assertTrue(value.isSet());
-        Assertions.assertNull(value.value());
-    }
-
-    @Test
-    void testGetNonnullPropertyValue()
-    {
-        final var value = propertyContainer.getPropertyValue(PROPERTY_STRING);
-
-        Assertions.assertNotNull(value);
-        Assertions.assertTrue(value.isSet());
-        Assertions.assertEquals(PROPERTY_STRING_DEFAULT, value.value());
-    }
-
-    @Test
-    void testGetUnsetProperty()
-    {
-        final var value = propertyContainer.getPropertyValue(PROPERTY_UNSET);
-
-        Assertions.assertNotNull(value);
-        Assertions.assertFalse(value.isSet());
-        Assertions.assertNull(value.value());
-    }
 
     @Test
     void testSetProperty()
@@ -120,44 +64,6 @@ class PropertyContainerTest
         propertyContainer.setPropertyValue(PROPERTY_UNSET, newValue);
         Assertions.assertEquals(newValue, propertyContainer.getPropertyValue(PROPERTY_UNSET).value());
         Assertions.assertTrue(propertyContainer.hasProperty(PROPERTY_UNSET));
-    }
-
-    @Test
-    void testHasProperty()
-    {
-        Assertions.assertTrue(propertyContainer.hasProperty(PROPERTY_STRING));
-
-        Assertions.assertFalse(propertyContainer.hasProperty(PROPERTY_UNSET));
-    }
-
-    @Test
-    void testHasProperties()
-    {
-        Assertions.assertTrue(propertyContainer.hasProperties(
-            PROPERTIES.get(0),
-            PROPERTIES.get(1)
-        ));
-
-        Assertions.assertFalse(propertyContainer.hasProperties(
-            PROPERTIES.get(0),
-            PROPERTY_UNSET
-        ));
-    }
-
-    @Test
-    void testGetRequiredPropertyValue()
-    {
-        final var quarterCircles = propertyContainer.getRequiredPropertyValue(PROPERTY_STRING);
-        Assertions.assertEquals(PROPERTY_STRING_DEFAULT, quarterCircles);
-    }
-
-    @Test
-    void testGetRequiredPropertyValueThrowsException()
-    {
-        Assertions.assertThrows(
-            NullPointerException.class,
-            () -> propertyContainer.getRequiredPropertyValue(PROPERTY_UNSET)
-        );
     }
 
     @Test
@@ -229,32 +135,6 @@ class PropertyContainerTest
 
         final var container = PropertyContainer.forType(mockStructureType);
         Assertions.assertEquals(propertyContainer, container);
-    }
-
-    @Test
-    void testIterator()
-    {
-        final var iterator = propertyContainer.iterator();
-
-        Assertions.assertTrue(iterator.hasNext());
-        var entry = iterator.next();
-        Assertions.assertEquals(PROPERTY_STRING.getFullKey(), entry.getKey());
-        Assertions.assertEquals(PROPERTY_STRING_DEFAULT, entry.getValue().value());
-
-        Assertions.assertTrue(iterator.hasNext());
-        entry = iterator.next();
-        Assertions.assertEquals(PROPERTY_NULLABLE.getFullKey(), entry.getKey());
-        Assertions.assertNull(entry.getValue().value());
-
-        Assertions.assertFalse(iterator.hasNext());
-    }
-
-    @Test
-    void testIteratorReadOnly()
-    {
-        final var iterator = propertyContainer.iterator();
-        iterator.next();
-        Assertions.assertThrows(UnsupportedOperationException.class, iterator::remove);
     }
 
     @Test
